fix: Only set signal handler when in main thread
As a safety precaution, pynisher will kill subprocesses
if the process get's a SIGTERM. This is done by setting
a signal handler with `signal.signal`. In the case where
this is for some reason not done on the main thread of
that process, then setting this handler will fail. While
it would be good to have a backup, this is likely not the case
and so we just don't set this signal handler if we
are not the main thread of the currently running
interpreter.